Fix update of exe_cnt upon calling db()
After removing the duplicated code between db() and db(),
exe_cnt was accedently was updated when it should have not.
During call for db(k!=0), the exe_cnt should not have been updated.
This didn't cause any issues since the calling for db(k!=0) was done
only once, when exe_cnt was equal to 0. Then, the exe_cnt added wqes
and made the management QP to send a doorbell index to the NIC allowing
the NIC to start the next allreduce, but only the recv_enable WQEs which
did not cause any issues.