Remove call for db(k) during CommGraph::finish()
No need to call the db(k) of the management QP right before
the CommGraph::finish() finishes because the execution of the
recv_enable WQEs can be executed in the begining of each run().

This was an optimization for the first run(), but all the sequentials
runs didn't benefit from it.

The correct way to do this is to send the recv_enable WQEs once during
the connect_and_prepare stage and not include them in the management QP
at all.